+++ src/core/dm_trans.c
@@ -52,6 +52,7 @@
 #include "device.h"
 #include "dm_inst_vector.h"
 #include "vendor_api.h"
+#include "text_utils.h"
 
 
 
@@ -143,11 +144,13 @@ void DM_TRANS_Add(dm_op_t op, char *path, char *value, dm_val_union_t *val_union
     int new_num_entries;
     dm_trans_t *dt;
     int i;
+    bool is_part_of_add;
 
     USP_ASSERT(cur_transaction != NULL);
 
-    // Do not add set operations, if they are part of a larger add operation - we only want to notify the add
+    // Determine whether the set operation was part of a larger add operation - we only want to notify the add
     // NOTE: When processing a USP AddRequest message, default values are not added to the transaction, only the overridden default values (in the USP AddRequest message)
+    is_part_of_add = false;
     if (op == kDMOp_Set)
     {
         // Iterate over all operations, seeing if any were an add operation
@@ -158,11 +161,12 @@ void DM_TRANS_Add(dm_op_t op, char *path, char *value, dm_val_union_t *val_union
             {
                 if (inst->order >= dt->inst.order)
                 {
-                    // If this set is for a parameter whose parent instances match that of the add, then do not add to transaction
+                    // If this set is for a parameter whose parent instances match that of the add, then it was part of a larger add operation
                     if ( (memcmp(inst->nodes, dt->node->instance_nodes, (dt->inst.order)*sizeof(dm_node_t *)) == 0) &&
                          (memcmp(inst->instances, dt->inst.instances, (dt->inst.order)*sizeof(int)) == 0) )
                     {
-                        return;
+                        is_part_of_add = true;
+                        break;
                     }
                 }
             }
@@ -178,6 +182,7 @@ void DM_TRANS_Add(dm_op_t op, char *path, char *value, dm_val_union_t *val_union
     dt->op = op;
     dt->path = USP_STRDUP(path);
     dt->node = node;
+    dt->is_part_of_add = is_part_of_add;
     memcpy(&dt->inst, inst, sizeof(dm_req_instances_t));
 
     // Save a copy of the value (if applicable to the operation being stored)
@@ -281,12 +286,15 @@ int DM_TRANS_Commit(void)
         {
             case kDMOp_Set:
                 // Internal notify callback
-                notify_set_cb = node->registered.param_info.notify_set_cb;
-                if (notify_set_cb != NULL)
+                if (dt->is_part_of_add == false)
                 {
-                    DM_PRIV_RequestInit(&req, node, dt->path, (dm_instances_t *) &dt->inst);
-                    req.val_union = dt->val_union;
-                    err = notify_set_cb(&req, dt->value);
+                    notify_set_cb = node->registered.param_info.notify_set_cb;
+                    if (notify_set_cb != NULL)
+                    {
+                        DM_PRIV_RequestInit(&req, node, dt->path, (dm_instances_t *) &dt->inst);
+                        req.val_union = dt->val_union;
+                        err = notify_set_cb(&req, dt->value);
+                    }
                 }
                 break;
 
@@ -450,6 +458,37 @@ bool DM_TRANS_IsWithinTransaction(void)
     return false;
 }
 
+/*********************************************************************//**
+**
+** DM_TRANS_GetParamWritesByPathSpec
+**
+** Returns a list of all parameters written in a transaction (matching a specified spec), pending being committed
+**
+** \param   path_spec - specification of the path to match (can contain, wildcards, instance number or partial path)
+** \param   kvv - pointer to key-value vector in which to return the list of parameters which match (and their values)
+**
+** \return  None
+**
+**************************************************************************/
+void DM_TRANS_GetParamWritesByPathSpec(char *path_spec, kv_vector_t *kvv)
+{
+    int i;
+    dm_trans_t *dt;
+    bool is_match;
+
+    for (i=0; i < cur_transaction->num_entries; i++)
+    {
+        dt = &cur_transaction->vector[i];
+        if (dt->op == kDMOp_Set)
+        {
+            is_match = TEXT_UTILS_IsPathMatch(dt->path, path_spec);
+            if (is_match)
+            {
+                KV_VECTOR_Add(kvv, dt->path, dt->value);
+            }
+        }
+    }
+}
 
 /*********************************************************************//**
 **